rename: authStore useAuthStore로 변경 (#139)
