+++ apps/web/src/app/(providers)/AuthProvider.tsx
@@ -1,7 +1,7 @@
 'use client';
 import type { ReactNode } from 'react';
 import { useEffect } from 'react';
-import { useAuthActions, useUserState } from 'src/entities/auth/stores/authStore';
+import { useAuthActions, useUserState } from 'src/entities/auth/stores/useAuthStore';
 import { createClient } from 'src/shared/supabase/client/client';
 import type { User } from '@supabase/supabase-js';
 
@@ -10,7 +10,7 @@ const AuthProvider = ({ user: initialUser, children }: { user: User | null; chil
   const { setUser, setLoading } = useAuthActions();
   const supabase = createClient();
 
-  useEffect(() => { 
+  useEffect(() => {
     if (initialUser && (!currentUser || currentUser.id !== initialUser.id)) {
       setUser(initialUser);
       setLoading(false);
