+++ apps/web/src/features/user/mypage/components/main/MyPageNavigation.tsx
@@ -3,7 +3,7 @@ import { FaGavel, FaCoins, FaHeart } from 'react-icons/fa';
 import { HiDocumentText } from 'react-icons/hi';
 import { HiChatBubbleLeftRight } from 'react-icons/hi2';
 import { LuHeartHandshake } from 'react-icons/lu';
-import { useUserState } from 'src/entities/auth/stores/authStore';
+import { useUserState } from 'src/entities/auth/stores/useAuthStore';
 import { ICON_NAMES, MYPAGE_MENU_LIST } from 'src/entities/user/mypage/main/constants';
 import MyPageNavigationList from 'src/features/user/mypage/components/main/MyPageNavigationList';
