+++ apps/web/src/shared/ui/AuthWrapper.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { notFound } from 'next/navigation';
-import { useUserLoadingState, useUserState } from 'src/entities/auth/stores/authStore';
+import { useUserLoadingState, useUserState } from 'src/entities/auth/stores/useAuthStore';
 
 const AuthWrapper = ({ children }: { children: React.ReactNode }) => {
   const user = useUserState();
